src/emailservice/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
backoff==1.10.0
    # via opentelemetry-exporter-otlp
cachetools==3.0.0
    # via google-auth
certifi==2018.11.29
    # via requests
chardet==3.0.4
    # via requests
google-api-core==1.6.0
    # via -r requirements.in
google-api-python-client==1.7.8
    # via google-cloud-profiler
google-auth-httplib2==0.0.3
    # via
    #   google-api-python-client
    #   google-cloud-profiler
google-auth==1.6.2
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-profiler
google-cloud-profiler==1.0.8
    # via -r requirements.in
googleapis-common-protos==1.52.0
    # via
    #   google-api-core
    #   opentelemetry-exporter-otlp
grpcio-health-checking==1.12.1
    # via -r requirements.in
grpcio==1.35.0
    # via
    #   -r requirements.in
    #   grpcio-health-checking
    #   opentelemetry-exporter-otlp
    #   opentelemetry-instrumentation-grpc
httplib2==0.18.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==2.8
    # via requests
jinja2==2.10.1
    # via
    #   -r requirements.in
    #   opentelemetry-instrumentation-jinja2
markupsafe==1.1.0
    # via jinja2
opentelemetry-api==0.16b1
    # via
    #   opentelemetry-exporter-otlp
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-grpc
    #   opentelemetry-instrumentation-jinja2
    #   opentelemetry-sdk
opentelemetry-exporter-otlp==0.16b1
    # via -r requirements.in
opentelemetry-instrumentation-grpc==0.16b1
    # via -r requirements.in
opentelemetry-instrumentation-jinja2==0.16b1
    # via -r requirements.in
opentelemetry-instrumentation==0.16b1
    # via
    #   -r requirements.in
    #   opentelemetry-instrumentation-jinja2
opentelemetry-proto==0.16b1
    # via opentelemetry-exporter-otlp
opentelemetry-sdk==0.16b1
    # via
    #   -r requirements.in
    #   opentelemetry-exporter-otlp
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-grpc
protobuf==3.14.0
    # via
    #   google-api-core
    #   google-cloud-profiler
    #   googleapis-common-protos
    #   grpcio-health-checking
    #   opentelemetry-proto
pyasn1-modules==0.2.3
    # via google-auth
pyasn1==0.4.5
    # via
    #   pyasn1-modules
    #   rsa
python-json-logger==0.1.9
    # via -r requirements.in
pytz==2018.9
    # via google-api-core
requests==2.21.0
    # via
    #   google-api-core
    #   google-cloud-profiler
rsa==4.0
    # via google-auth
six==1.12.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth
    #   grpcio
    #   protobuf
uritemplate==3.0.0
    # via google-api-python-client
urllib3==1.24.2
    # via requests
wrapt==1.12.1
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-jinja2

# The following packages are considered to be unsafe in a requirements file:
# setuptools
